Isolate UDP application fixtures from the process-global logger
